Mise en place d'une fixed time step. Désactivation de la limitation CPU.

+++ Code/ska_core/ECS/Basics/Input/InputComponent.h
@@ -5,7 +5,7 @@ namespace ska {
 	public:
 	    InputComponent();
 	    ~InputComponent() = default;
-		unsigned int movePower;
-		unsigned int jumpPower;
+		float movePower;
+		float jumpPower;
 	};
 }
